AgenteSFTPBackend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
PyJWT>=2.8.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
psycopg[binary]>=3.2.0,<4.0.0
pyodbc>=5.1.0,<6.0.0
